requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
numpy
nltk
tqdm
PyYAML
matplotlib
torch
notebook
ipykernel
boto3
mypy-boto3-s3
python-box
ensure
fastapi
uvicorn
Jinja2
gdown
-e .

